requirements.txt
nextcord.py
python-dotenv
pillow